Desktop/mcp-servers/bigquery-mcp-servers/requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0

mcp==1.26.0

google-cloud-bigquery==3.40.1
google-cloud-secret-manager==2.26.0

pydantic-settings==2.13.1
python-dotenv==1.0.1